Take into use e-mail address from Turku-Suomi.fi
This supposedly should have been in place from the start, but for some
reason it isn't. One possible reason is that the e-mail address Suomi.fi
gives us is not validated in any way - it could be garbage, or worse,
someone else's e-mail address. This change could have surprising side
effects because of that, but Varaamo does not work now if they don't get
the e-mail address.